make richEditor into component
